telegram_bot/requirements.txt
loguru==0.7.0
fastapi==0.109.2
python-telegram-bot==22.0
telegram==0.0.1
httpx==0.28.1
pydantic==2.11.3
pydantic_core==2.33.1
PyYAML==6.0.2
uvicorn==0.22.0